+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Enter here all the changes made to the development version
 
+### Fixed
+
+- Display No Review Text in Vendor Reviews Option Tabs
+
 
 ## [1.0.7] - 2020-10-28